Pull out cache and args module
