+++ src/args.rs
@@ -1,0 +1,10 @@
+// store some common string for argument names
+pub const DATA_DIR_ARG: &str = "data_dir";
+pub const FEATURES_ARG: &str = "supported_features";
+pub const CACHE_ARG: &str = "cache";
+pub const WASM_ARG: &str = "wasm";
+pub const CODE_ID_ARG: &str = "code_id";
+pub const MSG_ARG: &str = "msg";
+pub const ENV_ARG: &str = "env";
+pub const INFO_ARG: &str = "message_info";
+pub const GAS_USED_ARG: &str = "gas_used";
