+++ src/lib.rs
@@ -1,4 +1,6 @@
 mod api;
+mod args;
+mod cache;
 mod db;
 mod error;
 mod gas_meter;
@@ -14,27 +16,15 @@ pub use querier::GoQuerier;
 
 use std::convert::TryInto;
 use std::panic::{catch_unwind, AssertUnwindSafe};
-use std::str::from_utf8;
 
-use crate::error::{clear_error, handle_c_error, set_error, Error};
 use cosmwasm_vm::{
-    call_handle_raw, call_init_raw, call_migrate_raw, call_query_raw, features_from_csv, Backend,
-    Cache, CacheOptions, Checksum, InstanceOptions, Size,
+    call_handle_raw, call_init_raw, call_migrate_raw, call_query_raw, Backend, Cache, Checksum,
+    InstanceOptions,
 };
 
-const MEMORY_CACHE_SIZE: Size = Size::mebi(500); // TODO: Make configurable
-
-#[repr(C)]
-pub struct cache_t {}
-
-fn to_cache(ptr: *mut cache_t) -> Option<&'static mut Cache<GoStorage, GoApi, GoQuerier>> {
-    if ptr.is_null() {
-        None
-    } else {
-        let c = unsafe { &mut *(ptr as *mut Cache<GoStorage, GoApi, GoQuerier>) };
-        Some(c)
-    }
-}
+use crate::args::{CACHE_ARG, CODE_ID_ARG, ENV_ARG, GAS_USED_ARG, INFO_ARG, MSG_ARG, WASM_ARG};
+use crate::cache::{cache_t, to_cache};
+use crate::error::{handle_c_error, Error};
 
 fn into_backend(db: DB, api: GoApi, querier: GoQuerier) -> Backend<GoStorage, GoApi, GoQuerier> {
     Backend {
@@ -42,58 +32,6 @@ fn into_backend(db: DB, api: GoApi, querier: GoQuerier) -> Backend<GoStorage, Go
         api,
         querier,
     }
-}
-
-#[no_mangle]
-pub extern "C" fn init_cache(
-    data_dir: Buffer,
-    supported_features: Buffer,
-    err: Option<&mut Buffer>,
-) -> *mut cache_t {
-    let r = catch_unwind(|| do_init_cache(data_dir, supported_features))
-        .unwrap_or_else(|_| Err(Error::panic()));
-    match r {
-        Ok(t) => {
-            clear_error();
-            t as *mut cache_t
-        }
-        Err(e) => {
-            set_error(e, err);
-            std::ptr::null_mut()
-        }
-    }
-}
-
-// store some common string for argument names
-static DATA_DIR_ARG: &str = "data_dir";
-static FEATURES_ARG: &str = "supported_features";
-static CACHE_ARG: &str = "cache";
-static WASM_ARG: &str = "wasm";
-static CODE_ID_ARG: &str = "code_id";
-static MSG_ARG: &str = "msg";
-static ENV_ARG: &str = "env";
-static INFO_ARG: &str = "message_info";
-static GAS_USED_ARG: &str = "gas_used";
-
-fn do_init_cache(
-    data_dir: Buffer,
-    supported_features: Buffer,
-) -> Result<*mut Cache<GoStorage, GoApi, GoQuerier>, Error> {
-    let dir = unsafe { data_dir.read() }.ok_or_else(|| Error::empty_arg(DATA_DIR_ARG))?;
-    let dir_str = String::from_utf8(dir.to_vec())?;
-    // parse the supported features
-    let features_bin =
-        unsafe { supported_features.read() }.ok_or_else(|| Error::empty_arg(FEATURES_ARG))?;
-    let features_str = from_utf8(features_bin)?;
-    let features = features_from_csv(features_str);
-    let options = CacheOptions {
-        base_dir: dir_str.into(),
-        supported_features: features,
-        memory_cache_size: MEMORY_CACHE_SIZE,
-    };
-    let cache = unsafe { Cache::new(options) }?;
-    let out = Box::new(cache);
-    Ok(Box::into_raw(out))
 }
 
 /// frees a cache reference
